Add basic support for collisions
Each character has a unique set of collisionPoints. For now they
must be specified, but eventually it will default to the corners,
unless the character is bigger than the map's tilesize. In which
case, more collision points need to be there.

After the character is moved (before the camera is), the game
checks each collision point to see if it is touching an
impassable tile on the map (also specified in the settings object);

If there is a collision, the character's move is undone and the
camera is told not to move. Then update redraws the map as it was.