Add default collision points for the character
The default collision points start always with the corners of the
character. If the character size is bigger than the map's tile
size, then more points are added along the sides. This prevents
the corners of the tiles from being clear (and allowing the
character to move) while the center should be colliding with an
impassable tile.